requirements.txt
Flask
gunicorn
ipykernel
Jinja2
Keras
opencv-python
tensorflow-cpu
h5py==3.1.0
joblib
numpy==1.19.2
pandas
python-dotenv
pytz
six==1.15.0
sklearn
threadpoolctl
Werkzeug
WTForms
keras
wheel
matplotlib